requirements.txt
asgiref==3.2.3
Django==3.0
django-ckeditor==5.8.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==6.2.1
pygame==1.9.6
pytz==2019.3
sqlparse==0.3.0
